pyproject.toml
[tool.poetry]
name = "music-links-converter"
version = "0.1.0"
description = ""
authors = ["lizaa-t <lizaa.t@ya.ru>"]
readme = "README.md"
packages = [{include = "music_links_converter"}]

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^20.4"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
